requirements.txt
asgiref==3.7.2
Django==4.2.5
Pillow==10.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
sqlparse==0.4.4
